Rebase liquibase with noosi refacto
Make DbSchemaUpdater a service and call it where needed

+++ jqm-all/jqm-cli/src/main/java/com/enioka/jqm/cli/CommandImportJobDef.java
@@ -28,7 +28,7 @@ class CommandImportJobDef extends CommandBase
         {
             try
             {
-                cnx.runSelectSingle("q_select_default", Integer.class);
+                cnx.runSelectSingle("q_select_default", Long.class);
             }
             catch (NoResultException e)
             {
